SMACK-286 fix unit test for newer ant versions
Some unit tests depend on SmackConfiguration being loaded, so that the
initializer was run and the providers where configured. We create this
condition now by setting "reloading=false" to the junit task and hope
that there was a unit test before us that loaded SmackConfiguration.

But we keep in mind that this is only a ugly workaround, as unit test
should work on their own and have no inter-dependencies,