Don't use IQReplyFilter in AccountManager
This is a follow up on 6caf2cbdb52 where IQReplyFilter was changed so
that getUser() must not return null. But this is the case in
AccountManager where no resource binding has taken place and the user is
therefore not set.

The fix is the same as in 6caf2cbdb52, instead of IQReplyFilter we
simply use a PacketIDFilter.