SASL Proxy Auth support
This adds the ability to provide a distinct authorization identifier for use
by SASL mechanisms. Not all SASL mechanisms support this operation, in
particular CRAM-MD5.

Both the javax and provided SASL implementations are extended, and an authzid
parameter added to the authenticate method. Alternate methods are provided to
avoid API change.

The authorization identifier is passed as a BareJid in order to assure the
correct form.