Add support for XEP-0221: Data Forms Media Element.
While I was working on XEP-0232 (Software Information),
I felt the need of XEP-0221 (Data Forms Media Element).
So I gave it a try.

As you will go through the code, you will realize that
I struggled a bit with these issues :

MediaType.java :
The type/subtypes are expected to be as per
IANA MIME Media Types Registry.
Following as per this example
https://xmpp.org/extensions/xep-0221.html#example-1
the URI type codec=speex doesn't have an entry inside
IANA MIME Media Types Registry,
so I guess the user should be able to provide textual type/subtypes
even if they are yet to be included/are not included in
MIME Media Types Registry.
I wish to know if there is any way to add all those type/subtype
without entering them as string inside this class.

URINode.java :
I think that it's constructor impementation is a bit unconventional.
Inside the toXML funtion,
xml.attribute doesn't work because it gets confused for APOS_ENCODE,
so I harcoded it for now.

Also this work is still in progress but any suggestions are welcome :)