MUC: cleanly join not yet existing MUCs
This is a refinement of the MUC feature check and RoomInfo retrieval on
join. If you try to join a MUC that doesn't exist, which is valid, the
disco#info will fail with an item-not-found error, so we need to check
the service domain whether it's a MUC service, and can continue joining.
In that case, the roomInfo will remain unpopulated.

Signed-off-by: Georg Lukas <georg@op-co.de>