[tcp] Ignore exceptions in reader thread if writer was terminated
If we do not ignore the exception, then users may receive an exception
via connectionClosedOnError() on connection termination. Those
exceptions are typically unwanted if they are caused e.g. because the
server does not send a closing stream tag.

We previously ignored exceptions in this case already, but that
behavior was changed with [1: 57961a8cc1f2]. This commit re-adds the
behavior.

1: 57961a8cc1f2df6ecc1afa8c4f8460794d8d2dce
   Remove SynchronizationPoint