MAM - added support for multiple versions of MAM
The versions of the MAM extension that a chat server supports are exposed as features.
Version 1 of the MAM is exposed as feature 'urn:xmpp:mam:1', and version 2 as feature 'urn:xmpp:mam:2'.

The v2 of the extension is largely identical to v1, except for a number of additional filtering options.
These filtering options are optional for clients, and are not implemented in the Smack implementation of MAM.

When we ignore the optional fields, the only difference between MAMv1 and MAMv2 is the use of a different
extension namespace (urn:xmpp:mam:2 instead of urn:xmpp:mam:1) and a different FORM_TYPE (again urn:xmpp:mam:2 instead of
urn:xmpp:mam:1).

Smack currently only supports MAMv2 with namespace urn:xmpp:mam:2.

This commit changes Smack to support multiple versions of MAM, and to do discovery of the version of MAM supported by
the archive.
The default behavior is to do no discovery, and to use MAMv2. The behavior and API is unchanged.
When MamManager.enableMamVersionDiscovery() is called, newly created MamManager instances will
discover which version of MAM is supported by the archive, and use the highest version supported.