Bookmark parsing should not require non-empty nick
XEP-0048 defines an optional nickname for a bookmarked conference room, using an element named 'nick'.

Some clients have been observed to include an empty 'nick' element (instead of no such element) when a nickname is not stored as part of the bookmark.

Smack should not fail to parse these elements. Without the change in this commit, it fails with the following exception:

```
org.jxmpp.stringprep.XmppStringprepException: Argument can't be the empty string
    at org.jxmpp.stringprep.XmppStringPrepUtil.throwIfEmptyString(XmppStringPrepUtil.java:131)
    at org.jxmpp.stringprep.XmppStringPrepUtil.resourceprep(XmppStringPrepUtil.java:101)
    at org.jxmpp.jid.parts.Resourcepart.from(Resourcepart.java:91)
    at org.jivesoftware.smackx.bookmarks.Bookmarks.getConferenceStorage(Bookmarks.java:285)
    ...
```