[muc] do not wait infinite for the processing of the reflected self-presence
In theory, using an infinite wait() here should be fine. Because if
the collector in the lines before receives the presence, then it
should by guranteed by construction that the listener will eventually
also process it. However, Boris reported a case where the wait() would
hang forever. The actual root cause is unknown.

This change would fix the infinite wait(), but could theoretically
cause errornous NoResponseExceptions (e.g., if the system is under
heavy load and threads compete for the CPU).