LowLevelIntegrationTest#get*Connection() ignores descriptors
Low level Integration test, as implemented by `AbstractSmackLowLevelIntegrationTest` execute for each ConnectionDescriptor that is configured in XmppConnectionManager. However, their `getConnectedConnection()` and `getUnconnectedConnection()` methods will use the default connection descriptor. This can result in multiple executions of the same test, that all use the same connection descriptor.

This commit addresses the issue by allowing the to-be-used descriptor to be set prior to each test execution.