Debugger: Add 'context'
This adds a static map to SmackDebugger (the abstract base class of all Debuggers). This map is intended to be used to
store key/value pairs, to be used by subclasses, to enrich debugging functionality.

An earlier experiment implemented the context as a ThreadLocal, rather than a static map. Given the asynchronous nature
of much of Smack's API, this proved to be not very useful.