[muc] Bookmarks.getConferenceStorage() don't parse empty <nick/>
For a conference a server may return empty <nick/> element.
This will cause an exception on parsing and the whole conference won't appear in a client.
A similar problem will occur if something wrong with the nick format.
The nick in a conference is optional thing, and we are safe to ignore any problems with it.
In this case a default account JID must be used.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>