[sint] Add connection pre- and postconditions
This adds verification that before and after a test execution of AbstractSmackIntegrationTest instances, the connections that are used are in the expected state.

When these connections are _not_ in such a state, then it is likely that one of the tests contains a bug that will affect execution of other tests. Therefor, a breaking IOException is thrown when that's detected.