[muc] Fix deadlock on join/leave cycle
There was a potential deadlock if leave() is invoked it will wait for
the leave response using a stanza collector while holding the MUCs
monitor. However, the leave response is also processed by a stanza
listener invoking userHasLeft() which also requires the monitor lock
currently held by the thread invoking leave(). The stanza listener
trying to invoke userHasLeft() will now also be blocked, because the
MUC's monitor lock is not available. However, this will, in turn,
prevent the stanza collector from seeing the leave response, and the
thread invoking leave() will continue to hold the MUC monitor because
the collector does not return the response.

We fix this by not invoking userHasLeft() when we are currently
leaving or destroying the MUC. This fix requires that we change log
core logic so that the collectors are invoked *after* the listeners.

The main problem here is that a MUC service may send a leave/destroy
"response" (actually a unavailable presence) even without us
requesting it. For example, when the MUC service is (temporarily) shut
down. Therefore, we need to handle these cases as well.

Also add an integration test that tests for this case.