[sint] Debugger shouldn't error on pre-existing dirs
The StandardSinttestDebugger attempts to create directories in which output is stored. This code throws an error when the directory was not created.

If the directory already exists, it is not _created_, and thus an error is thrown.

With this code change, a directory creation attempt is performed only when the directory does not exist.