[sinttest] Low-level tests require account registration
Low-level tests are expected to leave the test account in an unexpected state. As such, they should not be executed with accounts that are reused for other tests.

When configuring the test framework with explicit accounts (e.g. with configuration like `-Dsinttest.accountOneUsername=jane -Dsinttest.accountOnePassword=secret`) such tests should not / cannot execute.

Prior to this commit, the above configuration would cause the entire test execution to error with `java.lang.IllegalStateException: Account registration not configured` if the run was configured to use explicit accounts and the test suite included low-level tests.

The change in this commit prevents that execution error, and instead marks low-level tests as 'impossible' when such configuration is detected.