[sinttest] add 'fail on impossible' configuration
Adds a new configuration option (`sinttest.failOnImpossibleTest`) that causes the test run process to exit with a non-zero exit code, if tests were impossible to run.

Test may be impossible to run, for example, because the server that's being tested does not support a particular feature.

It is not unreasonable for users to assume that, upon a successful test execution result, all tests have passed. In the case of 'impossible' tests, this isn't necessarily the case: some test might not have executed at all.

Especially in scenarios where users have configured the test run to enable a specific subset of the tests (through configuration such as `enabledTests` and `enabledSpecifications`) it may be desirable to hard fail when a test was impossible to execute. This enforces that a test run was strictly successful in all tests it was configured to execute.