[sinttest] Couple AccountManager's security config to SecurityMode
When the integration tests are configured without an administrative account (that creates test accounts) or explicit test accounts, the application will attempt to use In-Band Registration to create test accounts.

The AccountManager implementation by default refuses to operate, when the connection is not encrypted. This causes the In-Band Registration method to provision test accounts to fail, when running with SecurityMode.disabled / without encryption.

When security is disabled by configuration, it stands to reason that the user wants to allow AccountManager to work over an insecure connection. In this commit, a change is applied that ties the configuration together.