[core] Filter sun.security.pkcs11 from OSGi metadata
The OSGi plugin is smart and detects the usage of sun.security.pkcs11,
even though, Smack only accesses it via reflection and performs a
fallback if this is not possible. As a result, the OSGi plugin
declares sun.security.pkcs11 as Import-Package, which is causing
issues for Smack consumers using OSGi.

Filter the sun.security.pkcs11 package.