[pubsub] Detection of Pub/Sub service based on required feature
To detect a Pub/Sub service, Smack asserts that a service that advertises a pubsub identity also advertises a feature.

The feature that was used is contentious: although it is used in examples of XEP-0060, it is not used in normative text (nor is it defined as a known feature with the XSF Registrar).

In this commit, that feature is replaced by the 'subscribe' feature, which is one of two features that are required by the XEP to be supported.

The 'subscribe' feature is used instead of the 'publish' feature, as there is ambiguity in XEP-0060 regarding the mandatory nature of 'publish'. Section 10 defines it as being mandatory, while 7.1.1 defines it as optional.

fixes SMACK-957